requirements.txt
tensorflow>=2.12.0
numpy>=1.23.0
pandas>=1.5.0
scikit-learn>=1.2.0
matplotlib>=3.6.0
opencv-python>=4.7.0
Pillow>=9.4.0
fastapi>=0.95.0
uvicorn[standard]>=0.21.0
python-multipart>=0.0.6
streamlit>=1.22.0
plotly>=5.14.0
pydicom>=2.3.0
albumentations>=1.3.0
scipy>=1.10.0
tqdm>=4.65.0
pydantic>=1.10.0
